backend/requirements.txt
asgiref==3.10.0
black==25.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pytokens==0.3.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
